Added Geofencing (beta) and changed Schema to now store Names of the Users
Made a file to have a working geofencing model onky when the app is in foreground state, not background. Made a new document in firebase called user-details to store name and phone number, displayed it on the map. Main.dart has just been called the Geofencing service, but it's commented out for obvious reasons